Extract enum to avoid circular reference
